pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.0.1",
    "langchain>=1.0.2",
    "langchain-anthropic>=1.0.0",
    "langchain-community>=0.4",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.1",
    "mcp>=1.19.0",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
]
